+++ src/main/java/gift/controller/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package gift.controller;
+package gift.controller.response;
 
 public class ProductResponse {
