+++ src/main/java/gift/domain/Product.java
@@ -1,6 +1,6 @@
 package gift.domain;
 
-import gift.controller.ProductResponse;
+import gift.controller.response.ProductResponse;
 
 public class Product {
