+++ src/test/java/gift/service/ProductServiceTest.java
@@ -7,7 +7,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 import static org.mockito.BDDMockito.willDoNothing;
 
-import gift.controller.ProductRequest;
+import gift.controller.request.ProductRequest;
 import gift.domain.Product;
 import gift.domain.ProductRepository;
 import gift.error.exception.ProductNotFoundException;
